Started working on random dungeon generation ->added sun light as parameter of world generation ->added manhattan heuristic for astar (used for random path generation)

+++ GameTest/Sources/include/net/NetMessageTypes.h
@@ -125,7 +125,7 @@ struct NetMessage_WorldInit : public NetMessage_WorldSync
 
     int world_id;
     int player_id;
-    int dayTime;
+    float dayTime;
 
     std::string worldGeneratorModel;
     int worldGeneratorSeed;
